Update buttonloading.md
The values for setting the variable "CreatingNewUser" were switched. The First assign in the beginning is to set "CreatingNewUser" to TRUE, so to trigger the loading. The Second assign at the end is to set "CreatingNewUser" to FALSE to stop the loading.
I deleted the text "This is so the spinner doesn't display." because I think it was confusing. Also the image in step 5 is incorrect.
More over, the spinner will always show. With the setting "ShowLoadingAndLabel = False" what will not show is the Label of the button.